# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "laoshiren-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devices/register": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Register Device",
        "operationId": "register_device_api_v1_devices_register_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}/push-token": {
      "put": {
        "tags": [
          "devices"
        ],
        "summary": "Upsert Push Token",
        "operationId": "upsert_push_token_api_v1_devices__device_id__push_token_put",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "devices"
        ],
        "summary": "Delete Push Token",
        "operationId": "delete_push_token_api_v1_devices__device_id__push_token_delete",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/automations": {
      "get": {
        "tags": [
          "automations"
        ],
        "summary": "List Automations",
        "operationId": "list_automations_api_v1_automations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutomationResponse"
                  },
                  "title": "Response List Automations Api V1 Automations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "automations"
        ],
        "summary": "Create Automation",
        "operationId": "create_automation_api_v1_automations_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAutomationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/automations/notifications": {
      "get": {
        "tags": [
          "automations"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_api_v1_automations_notifications_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  },
                  "title": "Response List Notifications Api V1 Automations Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/automations/{automation_id}": {
      "get": {
        "tags": [
          "automations"
        ],
        "summary": "Get Automation",
        "operationId": "get_automation_api_v1_automations__automation_id__get",
        "parameters": [
          {
            "name": "automation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Automation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "automations"
        ],
        "summary": "Change Automation",
        "operationId": "change_automation_api_v1_automations__automation_id__patch",
        "parameters": [
          {
            "name": "automation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Automation Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAutomationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "automations"
        ],
        "summary": "Cancel Automation",
        "operationId": "cancel_automation_api_v1_automations__automation_id__delete",
        "parameters": [
          {
            "name": "automation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Automation Id"
            }
          },
          {
            "name": "expected_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Version"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutomationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "automation_type": {
            "$ref": "#/components/schemas/AutomationType"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "timezone_name": {
            "type": "string",
            "title": "Timezone Name"
          },
          "next_trigger_at": {
            "type": "string",
            "format": "date-time",
            "title": "Next Trigger At"
          },
          "thing_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thing Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "recurrence_interval_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Interval Seconds"
          },
          "status": {
            "$ref": "#/components/schemas/AutomationStatus"
          },
          "last_triggered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Triggered At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "automation_type",
          "title",
          "message",
          "timezone_name",
          "next_trigger_at",
          "thing_id",
          "task_id",
          "source_id",
          "recurrence_interval_seconds",
          "status",
          "last_triggered_at",
          "version",
          "created_at",
          "updated_at",
          "replayed"
        ],
        "title": "AutomationResponse"
      },
      "AutomationStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "PAUSED",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "AutomationStatus"
      },
      "AutomationType": {
        "type": "string",
        "enum": [
          "ONE_SHOT",
          "ONCE",
          "RECURRING",
          "RELATIVE",
          "CONDITION_WATCH",
          "CONDITION"
        ],
        "title": "AutomationType"
      },
...
      "ChangeAutomationRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "PAUSE",
              "RESUME",
              "CANCEL"
            ],
            "title": "Action"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "action",
          "expected_version"
        ],
        "title": "ChangeAutomationRequest"
      },
      "CreateAutomationRequest": {
        "properties": {
          "automation_type": {
            "$ref": "#/components/schemas/AutomationType"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          },
          "timezone": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Timezone"
          },
          "next_trigger_at": {
            "type": "string",
            "format": "date-time",
            "title": "Next Trigger At"
          },
          "thing_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thing Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "recurrence_interval_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 60.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Interval Seconds"
          }
        },
        "type": "object",
        "required": [
          "automation_type",
          "title",
          "message",
          "timezone",
          "next_trigger_at"
        ],
        "title": "CreateAutomationRequest"
      },
...
      "DevicePlatform": {
        "type": "string",
        "enum": [
          "HARMONYOS",
          "OTHER"
        ],
        "title": "DevicePlatform"
      },
      "DeviceRegisterRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "timezone": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Timezone"
          },
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform",
            "default": "HARMONYOS"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "timezone"
        ],
        "title": "DeviceRegisterRequest"
      },
      "DeviceResponse": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "platform",
          "timezone",
          "active",
          "last_seen_at"
        ],
        "title": "DeviceResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "automation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Automation Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "thing_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thing Id"
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "automation_id",
          "title",
          "message",
          "thing_id",
          "status",
          "attempt_count",
          "submitted_at",
          "error_code",
          "next_attempt_at",
          "created_at"
        ],
        "title": "NotificationResponse"
      },
      "NotificationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "SUBMITTED_TO_ADAPTER",
          "FAILED"
        ],
        "title": "NotificationStatus"
      },
...
      "PushTokenRequest": {
        "properties": {
          "push_token": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Push Token"
          },
          "provider": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Provider",
            "default": "HUAWEI"
          }
        },
        "type": "object",
        "required": [
          "push_token"
        ],
        "title": "PushTokenRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}